+++ dsh_sdk/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Changed
+- `read_access()` and `write_access()` are deprecated in favor of `is_readable()` and `is_writable()`, which provide clearer semantics.
+### Fixed
+- Return librdkafka compatible anchored read pattern in case of regex wildard
+- verify_list_of_topics now properly checks write topics including tenant name
+
 ## [0.8.1] - 2026-4-08
 ### Added
 - HTTP Protocol Adapter client